[core] move flare to utils/
- create a directory utils
- move configcheck and flare into utils
- address various comments #1422
- add pip freeze output
